+++ fleche/memo.ml
@@ -107,12 +107,12 @@ let interp_command ~st stm : _ Stats.t =
   let stm_loc = Coq.Ast.loc stm |> Option.get in
   match in_cache st stm with
   | Some (cached_loc, res), time ->
-    if Debug.cache then Io.Log.trace "memo" "cache hit";
+    if !Config.v.debug.cache then Io.Log.trace "memo" "cache hit";
     CacheStats.hit ();
     let res = adjust_offset ~stm_loc ~cached_loc res in
     Stats.make ~cache_hit:true ~time res
   | None, time_hash -> (
-    if Debug.cache then Io.Log.trace "memo" "cache miss";
+    if !Config.v.debug.cache then Io.Log.trace "memo" "cache miss";
     CacheStats.miss ();
     let kind = CS.Kind.Exec in
     let res, time_interp = CS.record ~kind ~f:(Coq.Interp.interp ~st) stm in